Adjust testing crates
