+++ testing/wasm-lightclient-tests/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wasm-tests"
+name = "wasm-lightclient-tests"
 version = "0.1.0"
 edition = "2021"
 publish = false
