+++ testing/wasm-rpc-tests/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "wasm-rpc-tests"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dev-dependencies]
+wasm-bindgen-test = "0.3.24"
+tracing-wasm = "0.2.1"
+tracing = "0.1.34"
+console_error_panic_hook = "0.1.7"
+serde_json = "1"
+futures-util = "0.3.28"
+
+# This crate is not a part of the workspace, because it
+# requires the "jsonrpsee web" features to be enabled, which we don't
+# want enabled for workspace builds in general.
+subxt = { path = "../../subxt", default-features = false, features = ["web", "jsonrpsee"] }
