+++ testing/wasm-rpc-tests/tests/wasm.rs
@@ -38,29 +38,3 @@ async fn wasm_ws_transport_works() {
     let chain = client.rpc().system_chain().await.unwrap();
     assert_eq!(&chain, "Development");
 }
-
-#[wasm_bindgen_test]
-async fn light_client_works() {
-    // Use a polkadot trusted DNS.
-    let api: LightClient<PolkadotConfig> = LightClientBuilder::new()
-        .bootnodes(
-            ["/dns/polkadot-connect-0.parity.io/tcp/443/wss/p2p/12D3KooWEPmjoRpDSUuiTjvyNDd8fejZ9eNWH5bE965nyBMDrB4o"]
-        )
-        .build_from_url("wss://rpc.polkadot.io:443")
-        .await
-        .expect("Cannot construct light client");
-
-    tracing::info!("Subscribe to latest finalized blocks: ");
-
-    let mut blocks_sub = api.blocks().subscribe_finalized().await.expect("Cannot subscribe to finalized hashes").take(3);
-    // For each block, print a bunch of information about it:
-    while let Some(block) = blocks_sub.next().await {
-        let block = block.expect("Block not valid");
-
-        let block_number = block.header().number;
-        let block_hash = block.hash();
-
-        tracing::info!("Block #{block_number}:");
-        tracing::info!("  Hash: {block_hash}");
-    }
-}
